Apply dependecy inversion principle by adding an abstract Storage class
